Added tests for BillController and BillService.

See merge request AdvProg/reguler-2023/mahasiswa/kelas-a/2106629963-Muflih-Naufal-Maxi-/kang-warnet/bayar-service!11

+++ src/main/java/id/ac/ui/cs/advprog/bayarservice/controller/BillController.java
@@ -24,7 +24,7 @@ public class BillController {
         );
     }
 
-    @PostMapping("invoices/{invoiceId}/bills")
+    @PostMapping("/invoices/{invoiceId}/bills")
     public ResponseEntity<Bill> addBillToInvoice(@PathVariable Integer invoiceId, @RequestBody BillRequest request) {
         Bill response = billService.create(request);
         return ResponseEntity.ok(response);
@@ -32,6 +32,9 @@ public class BillController {
 
     @DeleteMapping("/bills/delete/{bill_id}")
     public ResponseEntity<String> deleteBillById(@PathVariable Integer bill_id) {
+        if (billService.findById(bill_id) == null) {
+            return ResponseEntity.notFound().build();
+        }
         billService.delete(bill_id);
         return ResponseEntity.ok(String.format("Deleted Bill with id %d", bill_id));
     }
